services/workflow-orchestrator/pyproject.toml
[project]
name = "workflow-orchestrator"
version = "1.0.0"
description = "Dapr Workflow orchestrator for dynamic workflow execution with child workflow support"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "dapr>=1.14.0",
    "dapr-ext-workflow>=1.14.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "cel-python>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "workflows", "activities"]
